% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nlsy27}
\alias{nlsy27}
\title{National Longitudinal Survey of Youth 1997}
\format{
A list with components:
\describe{
\item{\code{dat}}{The response data}
\item{\code{Q}}{Intial design matrix with three factors and two to three specified loadings per factor}
\item{\code{cati}}{Indices of categorical (polytomous) items}
}
}
\usage{
nlsy27
}
\description{
A data set consisted of 3,458 individual responses to 27 mixed-type items, with a 1.12 percentage of missing data
}
\keyword{datasets}
