% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim18cfa0}
\alias{sim18cfa0}
\title{Simulated CFA data with LI}
\format{
A list with components:
\describe{
\item{\code{dat}}{A dataset with simulated responses of 1000 individuals to 18 items}
\item{\code{qlam}}{Loading pattern and values used to simulated the data}
}
}
\usage{
sim18cfa0
}
\description{
CFA data simulated based on 18 items, 3 factors and local independence;
 factorial correlation \eqn{\Phi=.3}.
}
\keyword{datasets}
