% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim18cfa1}
\alias{sim18cfa1}
\title{Simulated CFA data with LD}
\format{
A list with components:
\describe{
\item{\code{dat}}{A dataset with simulated responses of 1000 individuals to 18 items}
\item{\code{qlam}}{Loading pattern and values used to simulated the data}
\item{\code{LD}}{Local dependence between items (LD effect = .3)}
}
}
\usage{
sim18cfa1
}
\description{
CFA data simulated based on 18 items, 3 factors and local dependence; factorial correlation \eqn{\Phi=.3}.
}
\keyword{datasets}
