% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim18mcfa41}
\alias{sim18mcfa41}
\title{Simulated MCFA data with LD and Missingness}
\format{
A list with components:
\describe{
\item{\code{dat}}{A dataset with simulated responses of 500 individuals to 18 items}
\item{\code{qlam}}{Loading pattern and values used to simulated the data}
\item{\code{LD}}{Local dependence between items (LD effect = .3)}
}
}
\usage{
sim18mcfa41
}
\description{
CFA data mixed with continuous and categorical responses simulated based on 3 factors,
12 4-category items, 6 continuous items, local dependence, and 10% missigness at random;
factorial correlation \eqn{\Phi=.3}.
}
\keyword{datasets}
