% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim24ccfa21}
\alias{sim24ccfa21}
\title{Simulated CCFA data (dichotomous) with LD and a minor factor/trait}
\format{
A list with components:
\describe{
\item{\code{dat}}{A dataset with simulated responses of 1000 individuals to 24 items}
\item{\code{qlam}}{Loading pattern and values used to simulated the data}
\item{\code{LD}}{Local dependence between items (LD effect = .3)}
}
}
\usage{
sim24ccfa21
}
\description{
Categorical CFA data simulated based on 24 items, 4 factors, 2 categories
 and local dependence; factorial correlation \eqn{\Phi=.3}.
 The last factor/trait is minor (measured by cross-loadings only).
}
\keyword{datasets}
