% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lvm.R
\name{sim_lvm}
\alias{sim_lvm}
\title{Simulating data with Latent Variable Modeling}
\usage{
sim_lvm(
  N = 1000,
  K = 3,
  ipf = 8,
  cpf = 2,
  lam = 0.7,
  lac = 0.3,
  phi = 0.5,
  ph1 = -1,
  ecr = 0,
  ome_out = FALSE,
  cati = NULL,
  noc = c(4),
  misp = 0,
  rseed = 333,
  necw = K,
  necb = K,
  add_ind = c(),
  add_la = 0.5,
  add_phi = 0,
  zero_it = 0
)
}
\arguments{
\item{N}{Sample size.}

\item{K}{Number of factors.}

\item{ipf}{Items per factor.}

\item{cpf}{Cross-loadings per factor.}

\item{lam}{Number of formal iterations for posterior sampling.}

\item{lac}{Number of iterations to update the sampling information.}

\item{phi}{Homogeneous correlations between any two factors.}

\item{ph1}{Correlation between factor 1 and 2 (if it's different from \code{phi}.}

\item{ecr}{Residual correlation (local dependence).}

\item{ome_out}{Output factor score \eqn{\ome} or not.}

\item{cati}{The set of categorical (polytomous) items in sequence number (i.e., 1 to \eqn{J});
\code{NULL} for no and -1 for all (default is \code{NULL}).}

\item{noc}{Number of categories for categorical items}

\item{misp}{Proportion of missingness.}

\item{rseed}{An integer for the random seed.}
}
\value{
An object of class \code{list} containing the data, loading, and factorial correlation matrix.
}
\description{
\code{sim_lvm} can simulate data based on factor analytic models
 with different loading patterns and residual covariance (local dependence) structures.
}
\examples{

# for continuous data with cross-loadings and local dependence effect .3
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.3,ecr=.3,rseed = 123)
head(out$dat)
out$MLA
out$ofd_ind

# for categorical data with cross-loadings .4
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.4,cati=-1,noc=4,rseed = 123)
head(out$dat)
out$MLA\
out$ofd_ind

\donttest{
dat <- out$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1
Q
m0 <- pcfa(dat = dat, cati = -1, Q = Q, LD = FALSE,burn = 3000, iter = 3000,verbose=TRUE)
summary(m0) # summarize basic information
summary(m0, what = 'lambda') #summarize significant loadings
summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format
}
}
