% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lvm.R
\name{sim_lvm}
\alias{sim_lvm}
\title{Simulating data with Latent Variable Modeling}
\usage{
sim_lvm(
  N = 1000,
  lam = 0.7,
  K = 3,
  J = 18,
  cpf = 0,
  lac = 0.3,
  phi = 0.3,
  ecr = 0,
  necw = K,
  necb = K,
  P = 0,
  phix = 0,
  b = 0,
  lam1 = 0,
  K1 = 0,
  J1 = 0,
  b1 = 0,
  phi1 = 0,
  ilvl = NULL,
  cati = NULL,
  noc = c(4),
  misp = 0,
  fac_score = FALSE,
  rseed = 333,
  digits = 4
)
}
\arguments{
\item{N}{Sample size.}

\item{lam}{Loading value (for major loadings) or matrix (\eqn{J \times K}).}

\item{K}{Number of factors (if \code{lam} is a value).}

\item{J}{Number of items (if \code{lam} is a value).}

\item{cpf}{Number of cross-loadings per factor (if \code{lam} is a value).}

\item{lac}{Cross-loading value (if \code{lam} is a value).}

\item{phi}{Factor correlation scalar or matrix, or error correlations (for MIMIC-type model).}

\item{ecr}{Error covariance (local dependence) value.}

\item{necw}{Number of within-factor local dependence.}

\item{necb}{Number of between-factor local dependence.}

\item{P}{Number of observable predictors (for MIMIC-type model).}

\item{phix}{Observable predictor correlation value or matrix (for MIMIC-type model).}

\item{b}{Coefficients of observable predictors (for MIMIC-type model), value or \eqn{K \times P}.}

\item{lam1}{Loading value (for major loadings) or matrix (\eqn{J1 \times K1}) for latent predictors (for MIMIC-type model).}

\item{K1}{Number of latent predictors (if \code{lam} is a value, for MIMIC-type model).}

\item{J1}{Number of items latent predictors (if \code{lam} is a value, for MIMIC-type model).}

\item{b1}{Coefficients of latent predictors (for MIMIC-type model), value or \eqn{K \times K1}}

\item{phi1}{Latent predictor correlation scalar or matrix (for MIMIC-type model).}

\item{ilvl}{Specified levels of all items (i.e., need to specify Item 1 to \eqn{J+P+J1});
Any value smaller than 2 is considered as continuous item.}

\item{cati}{The set of polytomous items in sequence number (i.e., can be any number set
in between 1 and \eqn{J+P}); \code{NULL} for no and -1 for all (if \code{ilvl=NULL}).}

\item{noc}{Number of levels for polytomous items.}

\item{misp}{Proportion of missingness.}

\item{fac_score}{Output factor score or not.}

\item{rseed}{An integer for the random seed.}

\item{digits}{Number of significant digits to print when printing numeric values.}
}
\value{
An object of class \code{list} containing the data, loadings, factor correlations,
local dependence, and other information. The data consists of J items for the factors,
P items for observable predictors, and J1 items for latent predictors.
}
\description{
\code{sim_lvm} can simulate data with continuous latent variables (factors)
and continuous or categorical observed variables, plus a MIMIC-type structure.
One can also include an error covariance (local dependence) structure. Categorical
observed variables are generated with latent continuous responses normally distributed
and equally spaced within [-3,3].
}
\examples{

# for continuous data with cross-loadings and local dependence effect .3
out <- sim_lvm(N=1000,K=3,J=18,lam = .7, lac=.3,ecr=.3)
summary(out$dat)
out$lam
out$loc_dep

# for categorical data with cross-loadings .4 and 10\% missingness
out <- sim_lvm(N=1000,K=3,J=18,lam = .7, lac=.4,cati=-1,noc=4,misp=.1)
summary(out$dat)
out$lam
out$loc_dep

}
