% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lvm.R
\name{sim_lvm}
\alias{sim_lvm}
\title{Simulating data with Latent Variable Modeling}
\usage{
sim_lvm(
  N = 1000,
  K = 3,
  ipf = 8,
  cpf = 2,
  lam = 0.7,
  lac = 0.3,
  phi = 0.5,
  ph1 = -1,
  ecr = 0.3,
  ome_out = F,
  cati = NULL,
  noc = c(4),
  misp = 0,
  rseed = 333,
  necw = K,
  necb = K,
  add_ind = c(),
  add_la = 0.5,
  add_phi = 0,
  zero_it = 0
)
}
\arguments{
\item{N}{Sample size.}

\item{K}{Number of factors.}

\item{ipf}{Items per factor.}

\item{cpf}{Cross-loadings per factor.}

\item{lam}{Number of formal iterations for posterior sampling.}

\item{lac}{Number of iterations to update the sampling information.}

\item{phi}{= 0.5, ph1 = -1, logical; for adaptive Lasso or not. The default is \code{FALSE}, which seems slightly stabler.}

\item{cati}{The set of categorical (polytomous) items in sequence number (i.e., 1 to \eqn{J});
\code{NULL} for no and -1 for all (default is \code{NULL}).}

\item{rseed}{An integer for the random seed.}

\item{LD}{logical; \code{TRUE} for allowing LD (model with LD or C-step).}

\item{missing}{Value for missing data (default is \code{NA}).}

\item{digits}{Number of significant digits to print when printing numeric values.}

\item{verbose}{logical; to display the sampling information every \code{update} or not.}
}
\value{
\code{pcfa} returns an object of class \code{pcfa}. It contains a lot of information about
the posteriors that can be summarized using \code{\link{summary.lawbl}}. The factorial eigenvalue
 can be plotted with \code{\link{plot_eigen}}.
}
\description{
\code{sim_lvm} can simulate data based on factor analytic models
 with different loading patterns and residual covariance (local dependence) structures.
}
\examples{
\donttest{
####################################
#  Example 1: Estimation with LI   #
####################################

dat <- sim18cfa1$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1
Q
m0 <- pcfa(dat = dat, Q = Q, LD = FALSE)
summary(m0) # summarize basic information
summary(m0, what = 'lambda') #summarize significant loadings
summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format
summary(m0, what = 'offpsx') #summarize significant LD terms

}
}
\references{
Chen, J., Guo, Z., Zhang, L., & Pan, J. (2020). A partially confirmatory approach to scale development
 with the Bayesian Lasso. \emph{Psychological Methods}.
 Advance online publication. DOI:10.1037/met0000293
}
